+++ packages/coinjoin/src/utils/roundUtils.ts
@@ -165,15 +165,15 @@ const getDataFromRounds = (rounds: Round[]) => {
  * - maxMiningFee: array => value in kvBytes
  */
 export const transformStatus = ({
-    coinJoinFeeRateMedians,
+    coinjoinFeeRateMedians,
     roundStates: rounds,
 }: CoinjoinStatus) => {
     const { allowedInputAmounts, coordinationFeeRate } = getDataFromRounds(rounds);
-    // coinJoinFeeRateMedians include an array of medians per day, week and month - we take the second (week) median as the recommended fee rate
-    const recommendedMedian = coinJoinFeeRateMedians[1];
+    // coinjoinFeeRateMedians include an array of medians per day, week and month - we take the second (week) median as the recommended fee rate
+    const recommendedMedian = coinjoinFeeRateMedians[1];
     // the value is converted from kvBytes (kilo virtual bytes) to vBytes (how the value is displayed in UI)
     const maxMiningFee = recommendedMedian
-        ? Math.round(coinJoinFeeRateMedians[1].medianFeeRate / 1000)
+        ? Math.round(coinjoinFeeRateMedians[1].medianFeeRate / 1000)
         : MAX_MINING_FEE;
 
     return {